chore: improve offline support structure
