+++ src/index.ts
@@ -28,7 +28,7 @@ export * from './thread_manager';
 export * from './token_manager';
 export * from './types';
 export * from './channel_manager';
-export * from './offline_support_api';
+export * from './offline-support';
 // Don't use * here, that can break module augmentation https://github.com/microsoft/TypeScript/issues/46617
 export type {
   CustomAttachmentData,
