+++ src/offline-support/offline_support_api.ts
@@ -1,369 +1,16 @@
+import type { APIErrorResponse, ChannelResponse, Event } from '../types';
+
 import type {
-  APIErrorResponse,
-  AppSettingsAPIResponse,
-  ChannelAPIResponse,
-  ChannelFilters,
-  ChannelMemberResponse,
-  ChannelResponse,
-  ChannelSort,
-  Event,
-  LocalMessage,
-  Message,
-  MessageResponse,
-  PollResponse,
-  ReactionFilters,
-  ReactionResponse,
-  ReactionSort,
-  ReadResponse,
+  OfflineDBApi,
+  OfflineDBState,
+  PendingTask,
+  PreparedBatchQueries,
 } from './types';
+import type { StreamChat } from '../client';
 import type { AxiosError } from 'axios';
-import type { StreamChat } from './client';
-import type { Channel } from './channel';
-import { StateStore } from './store';
-import { runDetached } from './utils';
-
-// TODO: Find a clever way to pass this in a different manner
-/**
- * Represents a single SQL query or a query with parameters.
- */
-export type PreparedBatchQueries =
-  | [string]
-  | [string, Array<unknown> | Array<Array<unknown>>];
-
-/**
- * Options to insert a reaction into a message.
- */
-export type InsertReactionType = {
-  /** Message to which the reaction is applied. */
-  message: MessageResponse | LocalMessage;
-  /** The reaction to insert. */
-  reaction: ReactionResponse;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to upsert channel IDs associated with a query.
- */
-export type UpsertCidsForQueryType = {
-  /** Array of channel IDs. */
-  cids: string[];
-  /** Optional filters for the channels. */
-  filters?: ChannelFilters;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-  /** Optional sorting applied to the channels. */
-  sort?: ChannelSort;
-};
-
-/**
- * Options to upsert multiple channels.
- */
-export type UpsertChannelsType = {
-  /** Array of channel API responses. */
-  channels: ChannelAPIResponse[];
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-  /** If true, marks that the latest messages are already set. */
-  isLatestMessagesSet?: boolean;
-};
-
-/**
- * Options to upsert application settings for a user.
- */
-export type UpsertAppSettingsType = {
-  /** App settings data. */
-  appSettings: AppSettingsAPIResponse;
-  /** ID of the user the settings belong to. */
-  userId: string;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Sync status information for a user.
- */
-export type UpsertUserSyncStatusType = {
-  /** ID of the user. */
-  userId: string;
-  /** ISO timestamp of the last sync. */
-  lastSyncedAt: string;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to upsert a poll.
- */
-export type UpsertPollType = {
-  /** Poll data to be stored. */
-  poll: PollResponse;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to upsert individual channel data.
- */
-export type UpsertChannelDataType = {
-  /** Channel data. */
-  channel: ChannelResponse;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to upsert read statuses for a channel.
- */
-export type UpsertReadsType = {
-  /** Channel ID. */
-  cid: string;
-  /** Array of read statuses. */
-  reads: ReadResponse[];
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to upsert multiple messages.
- */
-export type UpsertMessagesType = {
-  /** Array of message responses. */
-  messages: MessageResponse[];
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to upsert members in a channel.
- */
-export type UpsertMembersType = {
-  /** Channel ID. */
-  cid: string;
-  /** Array of channel members. */
-  members: ChannelMemberResponse[];
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to update a reaction.
- */
-export type UpdateReactionType = {
-  /** Message associated with the reaction. */
-  message: MessageResponse | LocalMessage;
-  /** The updated reaction. */
-  reaction: ReactionResponse;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to update a message.
- */
-export type UpdateMessageType = {
-  /** Message to update. */
-  message: MessageResponse | LocalMessage;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to get channels by their IDs.
- */
-export type GetChannelsType = {
-  /** Array of channel IDs. */
-  cids: string[];
-  /** ID of the user. */
-  userId: string;
-};
-
-/**
- * Options to get channels based on filters.
- */
-export type GetChannelsForQueryType = {
-  /** ID of the user. */
-  userId: string;
-  /** Optional filters for channels. */
-  filters?: ChannelFilters;
-  /** Optional sorting for the channels. */
-  sort?: ChannelSort;
-};
-
-/**
- * Get the last sync timestamp for a user.
- */
-export type GetLastSyncedAtType = {
-  /** ID of the user. */
-  userId: string;
-};
-
-/**
- * Options to fetch pending tasks for a specific message.
- */
-export type GetPendingTasksType = {
-  /** Optional message ID to filter tasks. */
-  messageId?: string;
-};
-
-/**
- * Get application settings for a user.
- */
-export type GetAppSettingsType = {
-  /** ID of the user. */
-  userId: string;
-};
-
-/**
- * Options to retrieve reactions for a message.
- */
-export type GetReactionsType = {
-  /** ID of the message. */
-  messageId: string;
-  /** Optional filter to apply to reactions. */
-  filters?: Pick<ReactionFilters, 'type'>;
-  /** Optional sorting for reactions. */
-  sort?: ReactionSort;
-  /** Optional maximum number of reactions to return. */
-  limit?: number;
-};
-
-/**
- * Delete a pending task by ID.
- */
-export type DeletePendingTaskType = {
-  /** ID of the pending task. */
-  id: number;
-};
-
-/**
- * Options to delete a reaction from a message.
- */
-export type DeleteReactionType = {
-  /** The reaction to delete. */
-  reaction: ReactionResponse;
-  /** Optional message associated with the reaction. */
-  message?: MessageResponse | LocalMessage;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to delete a channel member.
- */
-export type DeleteMemberType = {
-  /** Channel ID. */
-  cid: string;
-  /** Member to remove. */
-  member: ChannelMemberResponse;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to drop all pending tasks for a message.
- */
-export type DropPendingTasksType = {
-  /** ID of the message. */
-  messageId: string;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to delete a message.
- */
-export type DeleteMessageType = {
-  /** ID of the message. */
-  id: string;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to delete a channel.
- */
-export type DeleteChannelType = {
-  /** Channel ID. */
-  cid: string;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Options to delete messages in a channel.
- */
-export type DeleteMessagesForChannelType = {
-  /** Channel ID. */
-  cid: string;
-  /** Timestamp before which messages are deleted. */
-  truncated_at?: string;
-  /** Whether to immediately execute the operation. */
-  execute?: boolean;
-};
-
-/**
- * Check if a channel exists by ID.
- */
-export type ChannelExistsType = {
-  /** Channel ID. */
-  cid: string;
-};
-
-/**
- * Represents a list of batch SQL queries to be executed.
- */
-export type ExecuteBatchQueriesType = PreparedBatchQueries[];
-
-export interface OfflineDBApi {
-  insertReaction: (options: InsertReactionType) => Promise<ExecuteBatchQueriesType>;
-  upsertCidsForQuery: (
-    options: UpsertCidsForQueryType,
-  ) => Promise<ExecuteBatchQueriesType>;
-  upsertChannels: (options: UpsertChannelsType) => Promise<ExecuteBatchQueriesType>;
-  upsertUserSyncStatus: (
-    options: UpsertUserSyncStatusType,
-  ) => Promise<ExecuteBatchQueriesType>;
-  upsertAppSettings: (options: UpsertAppSettingsType) => Promise<ExecuteBatchQueriesType>;
-  upsertPoll: (options: UpsertPollType) => Promise<ExecuteBatchQueriesType>;
-  upsertChannelData: (options: UpsertChannelDataType) => Promise<ExecuteBatchQueriesType>;
-  upsertReads: (options: UpsertReadsType) => Promise<ExecuteBatchQueriesType>;
-  upsertMessages: (options: UpsertMessagesType) => Promise<ExecuteBatchQueriesType>;
-  upsertMembers: (options: UpsertMembersType) => Promise<ExecuteBatchQueriesType>;
-  updateReaction: (options: UpdateReactionType) => Promise<ExecuteBatchQueriesType>;
-  updateMessage: (options: UpdateMessageType) => Promise<ExecuteBatchQueriesType>;
-  getChannels: (
-    options: GetChannelsType,
-  ) => Promise<Omit<ChannelAPIResponse, 'duration'>[] | null>;
-  getChannelsForQuery: (
-    options: GetChannelsForQueryType,
-  ) => Promise<Omit<ChannelAPIResponse, 'duration'>[] | null>;
-  getAllChannelCids: () => Promise<string[]>;
-  getLastSyncedAt: (options: GetLastSyncedAtType) => Promise<string | undefined>;
-  getAppSettings: (options: GetAppSettingsType) => Promise<AppSettingsAPIResponse | null>;
-  getReactions: (options: GetReactionsType) => Promise<ReactionResponse[] | null>;
-  executeSqlBatch: (queries: ExecuteBatchQueriesType) => Promise<unknown>;
-  addPendingTask: (task: PendingTask) => Promise<() => Promise<void>>;
-  getPendingTasks: (conditions?: GetPendingTasksType) => Promise<PendingTask[]>;
-  deletePendingTask: (options: DeletePendingTaskType) => Promise<ExecuteBatchQueriesType>;
-  deleteReaction: (options: DeleteReactionType) => Promise<ExecuteBatchQueriesType>;
-  deleteMember: (options: DeleteMemberType) => Promise<ExecuteBatchQueriesType>;
-  deleteChannel: (options: DeleteChannelType) => Promise<ExecuteBatchQueriesType>;
-  deleteMessagesForChannel: (
-    options: DeleteMessagesForChannelType,
-  ) => Promise<ExecuteBatchQueriesType>;
-  dropPendingTasks: (options: DropPendingTasksType) => Promise<ExecuteBatchQueriesType>;
-  hardDeleteMessage: (options: DeleteMessageType) => Promise<ExecuteBatchQueriesType>;
-  softDeleteMessage: (options: DeleteMessageType) => Promise<ExecuteBatchQueriesType>;
-  resetDB: () => Promise<unknown>;
-  channelExists: (options: ChannelExistsType) => Promise<boolean>;
-  initializeDB: () => Promise<boolean>;
-}
-
-export type OfflineDBState = {
-  initialized: boolean;
-  userId?: string;
-};
+import { OfflineDBSyncManager } from './offline_sync_manager';
+import { StateStore } from '../store';
+import { runDetached } from '../utils';
 
 /**
  * Abstract base class for an offline database implementation used with StreamChat.
@@ -1453,228 +1100,5 @@ export abstract class AbstractOfflineDB implements OfflineDBApi {
         id: task.id,
       });
     }
-  };
-}
-
-export type PendingTaskTypes = {
-  deleteMessage: 'delete-message';
-  deleteReaction: 'delete-reaction';
-  sendReaction: 'send-reaction';
-  sendMessage: 'send-message';
-};
-
-// TODO: Please rethink the definition of PendingTasks as it seems awkward
-export type PendingTask = {
-  channelId?: string;
-  channelType?: string;
-  messageId: string;
-  id?: number;
-} & (
-  | {
-      payload: Parameters<Channel['sendReaction']>;
-      type: PendingTaskTypes['sendReaction'];
-    }
-  | {
-      payload: Parameters<StreamChat['deleteMessage']>;
-      type: PendingTaskTypes['deleteMessage'];
-    }
-  | {
-      payload: Parameters<Channel['deleteReaction']>;
-      type: PendingTaskTypes['deleteReaction'];
-    }
-  | {
-      payload: Parameters<Channel['sendMessage']>;
-      type: PendingTaskTypes['sendMessage'];
-    }
-);
-
-export type PendingTaskExtraData = {
-  message?: Message;
-};
-
-/**
- * Manages synchronization between the local offline database and the Stream backend.
- *
- * Responsible for detecting connection changes, syncing channel data, and executing
- * pending tasks queued during offline periods. This class ensures the database remains
- * consistent with the server once connectivity is restored.
- */
-export class OfflineDBSyncManager {
-  public syncStatus = false;
-  public connectionChangedListener: { unsubscribe: () => void } | null = null;
-  private syncStatusListeners: Array<(status: boolean) => void> = [];
-  private scheduledSyncStatusCallbacks: Map<string | symbol, () => Promise<void>> =
-    new Map();
-  private client: StreamChat;
-  private offlineDb: AbstractOfflineDB;
-
-  constructor({
-    client,
-    offlineDb,
-  }: {
-    client: StreamChat;
-    offlineDb: AbstractOfflineDB;
-  }) {
-    this.client = client;
-    this.offlineDb = offlineDb;
-  }
-
-  /**
-   * Initializes the sync manager. Should only be called once per session.
-   *
-   * Cleans up old listeners if re-initialized to avoid memory leaks.
-   * Starts syncing immediately if already connected, otherwise waits for reconnection.
-   */
-  public init = async () => {
-    try {
-      // If the websocket connection is already active, then call
-      // the sync api straight away and also execute pending api calls.
-      // Otherwise wait for the `connection.changed` event.
-      if (this.client.user?.id && this.client.wsConnection?.isHealthy) {
-        await this.syncAndExecutePendingTasks();
-        await this.invokeSyncStatusListeners(true);
-      }
-
-      // If a listener has already been registered, unsubscribe from it so
-      // that it can be reinstated. This can happen if we reconnect with a
-      // different user or the component invoking the init() function gets
-      // unmounted and then remounted again. This part of the code makes
-      // sure the stale listener doesn't produce a memory leak.
-      if (this.connectionChangedListener) {
-        this.connectionChangedListener.unsubscribe();
-      }
-
-      this.connectionChangedListener = this.client.on(
-        'connection.changed',
-        async (event) => {
-          if (event.online) {
-            await this.syncAndExecutePendingTasks();
-            await this.invokeSyncStatusListeners(true);
-          } else {
-            await this.invokeSyncStatusListeners(false);
-          }
-        },
-      );
-    } catch (error) {
-      console.log('Error in DBSyncManager.init: ', error);
-    }
-  };
-
-  /**
-   * Registers a listener that is called whenever the sync status changes.
-   *
-   * @param listener - A callback invoked with the new sync status (`true` or `false`).
-   * @returns An object with an `unsubscribe` function to remove the listener.
-   */
-  public onSyncStatusChange = (listener: (status: boolean) => void) => {
-    this.syncStatusListeners.push(listener);
-
-    return {
-      unsubscribe: () => {
-        this.syncStatusListeners = this.syncStatusListeners.filter(
-          (el) => el !== listener,
-        );
-      },
-    };
-  };
-
-  /**
-   * Schedules a one-time callback to be invoked after the next successful sync.
-   *
-   * @param tag - A unique key to identify and manage the callback.
-   * @param callback - An async function to run after sync.
-   */
-  public scheduleSyncStatusChangeCallback = (
-    tag: string | symbol,
-    callback: () => Promise<void>,
-  ) => {
-    this.scheduledSyncStatusCallbacks.set(tag, callback);
-  };
-
-  /**
-   * Invokes all registered sync status listeners and executes any scheduled sync callbacks.
-   *
-   * @param status - The new sync status (`true` or `false`).
-   */
-  private invokeSyncStatusListeners = async (status: boolean) => {
-    this.syncStatus = status;
-    this.syncStatusListeners.forEach((l) => l(status));
-
-    if (status) {
-      const promises = Array.from(this.scheduledSyncStatusCallbacks.values()).map((cb) =>
-        cb(),
-      );
-      await Promise.all(promises);
-
-      this.scheduledSyncStatusCallbacks.clear();
-    }
-  };
-
-  /**
-   * Performs synchronization with the Stream backend.
-   *
-   * This includes downloading events since the last sync, updating the local DB,
-   * and handling sync failures (e.g., if syncing beyond the allowed retention window).
-   */
-  private sync = async () => {
-    if (!this.client?.user) {
-      return;
-    }
-    const cids = await this.offlineDb.getAllChannelCids();
-    // If there are no channels, then there is no need to sync.
-    if (cids.length === 0) {
-      return;
-    }
-
-    // TODO: We should not need our own user ID in the API, it can be inferred
-    const lastSyncedAt = await this.offlineDb.getLastSyncedAt({
-      userId: this.client.user.id,
-    });
-
-    if (lastSyncedAt) {
-      const lastSyncedAtDate = new Date(lastSyncedAt);
-      const nowDate = new Date();
-
-      // Calculate the difference in days
-      const diff = Math.floor(
-        (nowDate.getTime() - lastSyncedAtDate.getTime()) / (1000 * 60 * 60 * 24),
-      );
-
-      if (diff > 30) {
-        // stream backend will send an error if we try to sync after 30 days.
-        // In that case reset the entire DB and start fresh.
-        await this.offlineDb.resetDB();
-      } else {
-        try {
-          const result = await this.client.sync(cids, lastSyncedAtDate.toISOString());
-          const queryPromises = result.events.map((event) =>
-            this.offlineDb.handleEvent({ event, execute: false }),
-          );
-          const queriesArray = await Promise.all(queryPromises);
-          const queries = queriesArray.flat() as ExecuteBatchQueriesType;
-
-          if (queries.length) {
-            await this.offlineDb.executeSqlBatch(queries);
-          }
-        } catch (e) {
-          console.log('An error has occurred while syncing the DB.', e);
-          // Error will be raised by the sync API if there are too many events.
-          // In that case reset the entire DB and start fresh.
-          await this.offlineDb.resetDB();
-        }
-      }
-    }
-    await this.offlineDb.upsertUserSyncStatus({
-      userId: this.client.user.id,
-      lastSyncedAt: new Date().toString(),
-    });
-  };
-
-  /**
-   * Executes any tasks that were queued while offline and then performs a sync.
-   */
-  private syncAndExecutePendingTasks = async () => {
-    await this.offlineDb.executePendingTasks();
-    await this.sync();
   };
 }
